홈화면 ssr성공
