--- src/pages/home/index.ts
@@ -1,1 +1,0 @@
-export { default } from '../../service/Home';
